added converter to weather list

+++ Weather/app/src/main/java/com/just/weatherforecast/data/db/entity/MainWeatherParameters.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class MainWeatherParameters(
     @SerializedName("temp")
     val temperatureC: Double,
-    val pressure: Int,
-    val humidity: Int,
+    val pressure: Double,
+    val humidity: Double,
     @SerializedName("temp_min")
     val temperatureMin: Double,
     @SerializedName("temp_max")
