+++ Weather/app/src/main/java/com/just/weatherforecast/data/db/entity/Rain.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Rain(
     @SerializedName("3h")
-    val volume: Int
+    val volume: Double
 )
