+++ Weather/app/src/main/java/com/just/weatherforecast/data/db/entity/Wind.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class Wind(
     val speed: Double,
     @SerializedName("deg")
-    val direction: Int
+    val direction: Double
 )
